1. Elevated terminal rows/cols to uint16 2. Some toolbar buffer length rollover checks (this is LLM slop, but might as well be super safe) 3. Fix main issue: padding on info line of statusbar limit length to lesser of width or 80 4. Startup sequence no longer attempts to snap terminal to 80 columns 5. Show terminal size in 'i' info screen

+++ src/system_config.h
@@ -75,8 +75,8 @@ typedef struct _system_config {
 
     /** @name Terminal Display Settings */
     /**@{*/
-    uint8_t terminal_ansi_rows;                  /**< Terminal rows for ANSI display */
-    uint8_t terminal_ansi_columns;               /**< Terminal columns for ANSI display */
+    uint16_t terminal_ansi_rows;                  /**< Terminal rows for ANSI display */
+    uint16_t terminal_ansi_columns;               /**< Terminal columns for ANSI display */
     uint32_t terminal_ansi_color;                /**< ANSI color scheme */
     uint32_t terminal_ansi_statusbar;            /**< Status bar display mode */
     bool terminal_hide_cursor;                   /**< Hide terminal cursor */
